requirements.txt
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
streamlit==1.17.0